+++ tests/testthat/setup.R
@@ -1,13 +1,6 @@
 has_tern_key <- function() {
   tryCatch(
-    {
-      key <- keyring::key_get(
-        service = "TERN_API_KEY",
-        keyring = "nert"
-      )
-
-      nzchar(key)
-    },
-    error = function(e) FALSE
+    nzchar(get_key()),
+    nert_no_key = function(cnd) FALSE
   )
 }